+++ src/main/java/com/tristankechlo/improvedvanilla/eventhandler/BiomeLoadingHandler.java
@@ -3,6 +3,7 @@ package com.tristankechlo.improvedvanilla.eventhandler;
 import com.tristankechlo.improvedvanilla.init.ConfiguredStructures;
 import com.tristankechlo.improvedvanilla.structures.ForgottenWellStructure;
 import com.tristankechlo.improvedvanilla.structures.JungleTempleStructure;
+import com.tristankechlo.improvedvanilla.structures.UnderGroundTempleStructure;
 
 import net.minecraftforge.event.world.BiomeLoadingEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
@@ -21,6 +22,11 @@ public class BiomeLoadingHandler {
 		// add jungle temple biomes
 		if (JungleTempleStructure.DEFAULT_BIOMES.contains(event.getName().toString())) {
 			event.getGeneration().getStructures().add(() -> ConfiguredStructures.CONFIGURED_JUNGLE_TEMPLE);
+		}
+
+		// add underground temple biomes
+		if (UnderGroundTempleStructure.DEFAULT_BIOMES.contains(event.getName().toString())) {
+			event.getGeneration().getStructures().add(() -> ConfiguredStructures.CONFIGURED_UNDERGROUND_TEMPLE);
 		}
 
 	}