+++ Changelog.md
@@ -3,3 +3,6 @@
 ### Version 1.16.5 - 1.5.0
  - add forgotten well structure
  - add jungle temple structure
+ - add underground temple structure
+ - moved config to normal config folder
+